Refactor(controls): retrofit edit controls to sideDialog
