+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/customcontrols/buttons/ControlJoystick.java
@@ -19,7 +19,7 @@ import net.kdt.pojavlaunch.customcontrols.ControlData;
 import net.kdt.pojavlaunch.customcontrols.ControlJoystickData;
 import net.kdt.pojavlaunch.customcontrols.ControlLayout;
 import net.kdt.pojavlaunch.customcontrols.gamepad.GamepadJoystick;
-import net.kdt.pojavlaunch.customcontrols.handleview.EditControlPopup;
+import net.kdt.pojavlaunch.customcontrols.handleview.EditControlSideDialog;
 
 import org.lwjgl.glfw.CallbackBridge;
 
@@ -121,7 +121,7 @@ public class ControlJoystick extends JoystickView implements ControlInterface {
     public void sendKeyPresses(boolean isDown) {/*STUB since non swipeable*/ }
 
     @Override
-    public void loadEditValues(EditControlPopup editControlPopup) {
+    public void loadEditValues(EditControlSideDialog editControlPopup) {
         editControlPopup.loadJoystickValues(mControlData);
     }
