+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/customcontrols/buttons/ControlSubButton.java
@@ -2,13 +2,12 @@ package net.kdt.pojavlaunch.customcontrols.buttons;
 
 import android.annotation.SuppressLint;
 import android.view.MotionEvent;
-import android.view.View;
 import android.view.ViewGroup;
 
 import net.kdt.pojavlaunch.customcontrols.ControlData;
 import net.kdt.pojavlaunch.customcontrols.ControlDrawerData;
 import net.kdt.pojavlaunch.customcontrols.ControlLayout;
-import net.kdt.pojavlaunch.customcontrols.handleview.EditControlPopup;
+import net.kdt.pojavlaunch.customcontrols.handleview.EditControlSideDialog;
 
 @SuppressLint("ViewConstructor")
 public class ControlSubButton extends ControlButton {
@@ -91,7 +90,7 @@ public class ControlSubButton extends ControlButton {
     }
 
     @Override
-    public void loadEditValues(EditControlPopup editControlPopup) {
+    public void loadEditValues(EditControlSideDialog editControlPopup) {
         editControlPopup.loadSubButtonValues(getProperties(), parentDrawer.drawerData.orientation);
     }
 }
